创建订单接口

+++ open-home-app/open-home-resource/src/main/java/com/shihui/openpf/home/resource/MerchantGoodsResource.java
@@ -66,7 +66,7 @@ public class MerchantGoodsResource {
     public String update(
             @Context RequestContext rc,
             @ParamDesc(desc = "商户id", isRequired = false) @FormParam("merchant_id") Integer merchant_id,
-            @ParamDesc(desc = "商品id", isRequired = false) @FormParam("goods_id") Integer goods_id,
+            @ParamDesc(desc = "商品id", isRequired = false) @FormParam("goods_id") Long goods_id,
             @ParamDesc(desc = "商户商品开通状态", isRequired = false) @FormParam("status") Integer status,
             @ParamDesc(desc = "商户商品开通状态", isRequired = false) @FormParam("settlement") String settlement
 
@@ -86,7 +86,7 @@ public class MerchantGoodsResource {
     public String create(
             @Context RequestContext rc,
             @ParamDesc(desc = "商户id", isRequired = true) @FormParam("merchant_id") Integer merchant_id,
-            @ParamDesc(desc = "商品id", isRequired = true) @FormParam("goods_id") Integer goods_id,
+            @ParamDesc(desc = "商品id", isRequired = true) @FormParam("goods_id") Long goods_id,
             @ParamDesc(desc = "分类id", isRequired = true) @FormParam("category_id") Integer category_id,
             @ParamDesc(desc = "业务id", isRequired = true) @FormParam("service_id") Integer service_id,
             @ParamDesc(desc = "商户商品开通状态", isRequired = true) @FormParam("status") Integer status,
